+++ docker/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the docker containers will be documented in this file.
 
+### 2026-07-06
+- Default to the Client Spider when -j (the modern option) specified.
+
 ### 2026-06-23
 - Added options to select the modern spider.